ы

+++ task/Car.cs
@@ -28,18 +28,18 @@ namespace task
         // Публичные поля
         public string Name { get; protected set; }
         public string Model { get; protected set; }
-        public byte Speed { get; protected set; }
+        public int Speed { get; protected set; }
         public string Graphics { get; protected set; }
-        public ushort TraveledDistance { get; protected set; }
-        public byte CurrentSpeed { get; protected set; }
+        public int TraveledDistance { get; protected set; }
+        public int CurrentSpeed { get; protected set; }
 
         // Поля для внутреннего использования.
-        protected byte _speedRange;
-        protected byte _accelerationSpeed;
+        protected int _speedRange;
+        protected int _accelerationSpeed;
         protected bool _reachedMaxSpeed;
 
         // Конструктор с параметрами.
-        public Car(string name, string model, byte speed)
+        public Car(string name, string model, int speed)
         {
             Name = name;
             Model = model;
@@ -48,7 +48,7 @@ namespace task
             CurrentSpeed = 0;
             _reachedMaxSpeed = false;
 
-            // подписка на события гонки: её начало и обновление.
+            // Подписка на события гонки: её начало и обновление.
             RacingGame.Instance.StartEvent += Start;
             RacingGame.Instance.UpdateEvent += Update;
         }
